implemented PlatformImpl, a triggerable platform

+++ src/main/java/it/unibo/elementsduo/resources/Position.java
@@ -1,4 +1,4 @@
 package it.unibo.elementsduo.resources;
 
-public record Position(int x, int y) {
+public record Position(double x, double y) {
 }
